// src/types/types.tsx
export interface NewQuestionProps {
    /**
     * a function to add a new question to the application
     * @param question 
     * @returns 
     */
    addQuestion: (question: {
      title: string;
      text: string;
      tags: string[];
      askedBy: string;
    }) => void;
    /**
     * A function to set the next page 
     * to be displayed after a new question is asked
     */
    handleQuestions?: () => void;
}